# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Publisher Author Marketplace API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/profiles/me": {
      "get": {
        "operationId": "profiles_me",
        "tags": ["profiles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "role": {
                          "type": "string",
                          "enum": ["author", "publisher"]
                        },
                        "displayName": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 120
                        },
                        "profilePhotoUrl": {
                          "type": "string",
                          "format": "uri",
                          "maxLength": 2048,
                          "nullable": true
                        },
                        "signupIntent": {},
                        "approvalStatus": {
                          "type": "string",
                          "enum": ["pending", "approved", "rejected"]
                        },
                        "eligibilityStatus": {
                          "type": "string",
                          "enum": [
                            "eligible",
                            "limited",
                            "blocked",
                            "quarantined"
                          ],
                          "default": "limited"
                        },
                        "reviewOutcome": {
                          "type": "string",
                          "enum": [
                            "auto_approved",
                            "needs_review",
                            "admin_approved",
                            "admin_rejected",
                            "quarantined"
                          ],
                          "default": "needs_review"
                        },
                        "locale": {
                          "type": "string",
                          "enum": ["tr", "en"]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "userId",
                        "role",
                        "displayName",
                        "profilePhotoUrl",
                        "signupIntent",
                        "approvalStatus",
                        "locale",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "details": {}
                  },
                  "required": ["profile", "details"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles": {
      "post": {
        "operationId": "profiles_create",
        "tags": ["profiles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "role": {
                          "type": "string",
                          "enum": ["author", "publisher"]
                        },
                        "displayName": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 120
                        },
                        "profilePhotoUrl": {
                          "type": "string",
                          "format": "uri",
                          "maxLength": 2048,
                          "nullable": true
                        },
                        "signupIntent": {},
                        "approvalStatus": {
                          "type": "string",
                          "enum": ["pending", "approved", "rejected"]
                        },
                        "eligibilityStatus": {
                          "type": "string",
                          "enum": [
                            "eligible",
                            "limited",
                            "blocked",
                            "quarantined"
                          ],
                          "default": "limited"
                        },
                        "reviewOutcome": {
                          "type": "string",
                          "enum": [
                            "auto_approved",
                            "needs_review",
                            "admin_approved",
                            "admin_rejected",
                            "quarantined"
                          ],
                          "default": "needs_review"
                        },
                        "locale": {
                          "type": "string",
                          "enum": ["tr", "en"]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "userId",
                        "role",
                        "displayName",
                        "profilePhotoUrl",
                        "signupIntent",
                        "approvalStatus",
                        "locale",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "details": {}
                  },
                  "required": ["profile", "details"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles/me/onboarding-details": {
      "post": {
        "operationId": "profiles_completeDetails",
        "tags": ["profiles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "role": {
                          "type": "string",
                          "enum": ["author", "publisher"]
                        },
                        "displayName": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 120
                        },
                        "profilePhotoUrl": {
                          "type": "string",
                          "format": "uri",
                          "maxLength": 2048,
                          "nullable": true
                        },
                        "signupIntent": {},
                        "approvalStatus": {
                          "type": "string",
                          "enum": ["pending", "approved", "rejected"]
                        },
                        "eligibilityStatus": {
                          "type": "string",
                          "enum": [
                            "eligible",
                            "limited",
                            "blocked",
                            "quarantined"
                          ],
                          "default": "limited"
                        },
                        "reviewOutcome": {
                          "type": "string",
                          "enum": [
                            "auto_approved",
                            "needs_review",
                            "admin_approved",
                            "admin_rejected",
                            "quarantined"
                          ],
                          "default": "needs_review"
                        },
                        "locale": {
                          "type": "string",
                          "enum": ["tr", "en"]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "userId",
                        "role",
                        "displayName",
                        "profilePhotoUrl",
                        "signupIntent",
                        "approvalStatus",
                        "locale",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "details": {}
                  },
                  "required": ["profile", "details"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profiles/me/match-visible-contacts": {
      "put": {
        "operationId": "profiles_updateMatchVisibleContacts",
        "tags": ["profiles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "publicEmail": {
                    "type": "string",
                    "maxLength": 254,
                    "nullable": true
                  },
                  "publicPhone": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 40,
                    "nullable": true
                  },
                  "websiteUrl": {},
                  "socialLinks": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "label": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 40
                        },
                        "url": {
                          "type": "string",
                          "format": "uri",
                          "maxLength": 2048
                        },
                        "visible": {
                          "type": "boolean",
                          "default": false
                        }
                      },
                      "required": ["label", "url"],
                      "additionalProperties": false
                    },
                    "maxItems": 8,
                    "default": []
                  },
                  "visibility": {
                    "type": "object",
                    "properties": {
                      "publicEmail": {
                        "type": "boolean",
                        "default": false
                      },
                      "publicPhone": {
                        "type": "boolean",
                        "default": false
                      },
                      "websiteUrl": {
                        "type": "boolean",
                        "default": false
                      },
                      "socialLinks": {
                        "type": "boolean",
                        "default": false
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "required": ["websiteUrl", "visibility"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": {
                      "type": "object",
                      "properties": {
                        "publicEmail": {
                          "type": "string",
                          "maxLength": 254,
                          "nullable": true
                        },
                        "publicPhone": {
                          "type": "string",
                          "minLength": 3,
                          "maxLength": 40,
                          "nullable": true
                        },
                        "websiteUrl": {},
                        "socialLinks": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "label": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 40
                              },
                              "url": {
                                "type": "string",
                                "format": "uri",
                                "maxLength": 2048
                              },
                              "visible": {
                                "type": "boolean",
                                "default": false
                              }
                            },
                            "required": ["label", "url"],
                            "additionalProperties": false
                          },
                          "maxItems": 8,
                          "default": []
                        },
                        "visibility": {
                          "type": "object",
                          "properties": {
                            "publicEmail": {
                              "type": "boolean",
                              "default": false
                            },
                            "publicPhone": {
                              "type": "boolean",
                              "default": false
                            },
                            "websiteUrl": {
                              "type": "boolean",
                              "default": false
                            },
                            "socialLinks": {
                              "type": "boolean",
                              "default": false
                            }
                          },
                          "additionalProperties": false
                        }
                      },
                      "required": ["websiteUrl", "visibility"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["settings"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/publishers": {
      "get": {
        "operationId": "profiles_publicPublishers",
        "tags": ["profiles"],
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "publishers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 120
                          },
                          "logoUrl": {},
                          "websiteUrl": {}
                        },
                        "required": ["id", "name", "logoUrl", "websiteUrl"],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["publishers"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-webhook-signature"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {}
            },
            "required": ["code", "message"],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false
      }
    }
  }
}